pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-study-notes"
version = "1.0.0"
description = "Bilingual Machine Learning Study Notes - Interview Ready"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "ML Study Notes Team"}
]
keywords = ["machine-learning", "study-notes", "interview-prep", "bilingual"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.21.0",
    "matplotlib>=3.5.0",
    "scikit-learn>=1.0.0",
    "torch>=1.12.0",
    "markdown>=3.4.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
ml-validate = "validators.main:main"
ml-site = "site_generator.build:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ml_examples*", "validators*", "site_generator*", "skills*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E501"]
